# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /rfqs:
    get:
      operationId: listRfqs
      tags: [rfqs]
      summary: List all RFQs
      responses:
        "200":
          description: All RFQs with embedded quotes
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/RfqWithQuotes" }
    post:
      operationId: createRfq
      tags: [rfqs]
      summary: Create a new RFQ
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/RfqCreate" }
      responses:
        "201":
          description: Created RFQ
          content:
            application/json:
              schema: { $ref: "#/components/schemas/RfqWithQuotes" }
  /rfqs/{id}:
    get:
      operationId: getRfq
      tags: [rfqs]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: RFQ with embedded quotes
          content:
            application/json:
              schema: { $ref: "#/components/schemas/RfqWithQuotes" }
        "404":
          description: Not found
    patch:
      operationId: updateRfq
      tags: [rfqs]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/RfqUpdate" }
      responses:
        "200":
          description: Updated RFQ
          content:
            application/json:
              schema: { $ref: "#/components/schemas/RfqWithQuotes" }
        "404":
          description: Not found
  /rfqs/{id}/quotes:
    post:
      operationId: addRfqQuote
      tags: [rfqs]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/RfqQuoteCreate" }
      responses:
        "201":
          description: Created quote
          content:
            application/json:
              schema: { $ref: "#/components/schemas/RfqQuote" }
        "404":
          description: RFQ not found
  /rfqs/{id}/quotes/{quoteId}:
    patch:
      operationId: updateRfqQuote
      tags: [rfqs]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
        - in: path
          name: quoteId
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/RfqQuoteUpdate" }
      responses:
        "200":
          description: Updated quote
          content:
            application/json:
              schema: { $ref: "#/components/schemas/RfqQuote" }
        "404":
          description: Not found
    delete:
      operationId: deleteRfqQuote
      tags: [rfqs]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
        - in: path
          name: quoteId
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Deleted
  /rfqs/{id}/convert:
    post:
      operationId: convertRfqToPo
      tags: [rfqs]
      summary: Accept winning quote and create a live shipment + deal
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/RfqConvertInput" }
      responses:
        "201":
          description: Newly created shipment
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Shipment" }
        "400":
          description: Validation error (no accepted quote, already converted, etc.)
        "404":
          description: RFQ not found
  /rfqs/{id}/proforma:
    get:
      operationId: getRfqProforma
      tags: [rfqs]
      summary: Get proforma invoice data for a converted RFQ
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Proforma invoice data
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ProformaInvoice" }
        "404":
          description: Not found or not yet converted
...
  /rfqs/{id}/send-email:
    post:
      operationId: sendRfqEmail
      tags: [rfqs]
      summary: Send the RFQ details to supplier contacts via email
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/SendRfqEmailBody" }
      responses:
        "200":
          description: Email sent successfully
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SendRfqEmailResponse" }
        "404":
          description: RFQ not found
        "500":
          description: Email delivery failed
...
components:
  schemas:
...
    Payment:
      type: object
      required: [id, shipmentId, label, percent, amountUsd, paid, dueDate, sortOrder]
      properties:
        id:                      { type: integer }
        shipmentId:              { type: integer }
        label:                   { type: string }
        percent:                 { type: integer }
        amountUsd:               { type: integer }
        paid:                    { type: boolean }
        dueDate:                 { type: string, format: date-time }
        sortOrder:               { type: integer }
        paidAt:                  { type: string, format: date-time, nullable: true }
        referenceNumber:         { type: string, nullable: true }
        method:                  { type: string, nullable: true }
        buyerSharePct:                  { type: integer, nullable: true, description: "Percentage of amountUsd the Buyer committed upfront (e.g. 40). Null means no intermediary financing on this payment." }
        intermediaryAdvanceUsd:         { type: integer, nullable: true, description: "USD amount the Intermediary has fronted to the Supplier on behalf of the Buyer." }
        intermediaryRecoveredUsd:       { type: integer, nullable: true, description: "USD amount the Intermediary has already recovered from the Buyer." }
        invoiceNumber:                  { type: string, nullable: true, description: "Supplier invoice number for this payment." }
        intermediarySupplierPaidUsd:    { type: integer, nullable: true, description: "USD amount the Intermediary has actually wired to the Supplier." }
        intermediarySupplierPaidAt:     { type: string, format: date-time, nullable: true, description: "Date the Intermediary wired funds to the Supplier." }
...
    FactoryQuote:
      type: object
      required: [id, shipmentId, factory, country, unitPrice, leadDays, moq, selected, sortOrder]
      properties:
        id:           { type: integer }
        shipmentId:   { type: integer }
        factory:      { type: string }
        country:      { type: string }
        unitPrice:    { type: number }
        leadDays:     { type: integer }
        moq:          { type: integer }
        selected:     { type: boolean }
        sortOrder:    { type: integer }
        validityDate: { type: string, nullable: true }
        notes:        { type: string, nullable: true }
...
    Shipment:
      type: object
      required:
        - id
        - poNumber
        - product
        - category
        - supplierId
        - supplierName
        - customerName
        - status
        - currentStageId
        - dueDate
        - exFactoryDate
        - destination
        - via
        - payments
        - quotes
      properties:
        id:             { type: integer }
        poNumber:        { type: string, description: "Supplier-facing PO number (trader→supplier)" }
        buyerPoNumber:   { type: ["string", "null"], description: "Primary buyer-facing PO number (buyer→trader). Null if no deal is linked." }
        buyerPoNumbers:
          type: array
          items: { type: string }
          description: "All buyer-facing PO numbers linked to this shipment via the deal_shipments join table."
        product:        { type: string }
        category:       { type: string }
        supplierId:     { type: integer }
        supplierName:   { type: string }
        customerName:   { type: string }
        dealId:         { type: ["integer", "null"] }
        status:         { type: string }
        currentStageId: { type: string }
        dueDate:        { type: string, format: date-time }
        exFactoryDate:  { type: string, format: date-time }
        destination:    { type: string }
        via:            { type: string }
        notes:          { type: ["string", "null"] }
        quantity:       { type: integer, nullable: true }
        unitCostUsd:    { type: integer, nullable: true }
        buyerUnitPrice: { type: ["number", "null"], description: "Buyer-facing unit price from the linked deal. Null if no deal is linked." }
        buyerQuantity:  { type: ["integer", "null"], description: "Buyer quantity from the linked deal. Null if no deal is linked." }
        spreadUsd:      { type: ["number", "null"], description: "Gross spread in USD (buyer total − supplier cost). Null if no deal is linked." }
        spreadPct:      { type: ["number", "null"], description: "Gross margin % (spreadUsd / buyerTotalUsd × 100). Null if no deal is linked or buyer total is zero." }
        assigneeId:     { type: ["string", "null"], description: "Clerk userId of the assigned team member" }
        assigneeName:   { type: ["string", "null"], description: "Display name of the assigned team member" }
        payments:
          type: array
          items: { $ref: "#/components/schemas/Payment" }
        quotes:
          type: array
          items: { $ref: "#/components/schemas/FactoryQuote" }
...
    RfqQuote:
      type: object
      required: [id, rfqId, factoryName, country, unitPriceUsd, leadTimeDays, moq, status, sortOrder]
      properties:
        id:            { type: integer }
        rfqId:         { type: integer }
        supplierId:    { type: ["integer", "null"] }
        factoryName:   { type: string }
        country:       { type: string }
        unitPriceUsd:  { type: number }
        leadTimeDays:  { type: integer }
        moq:           { type: integer }
        notes:         { type: ["string", "null"] }
        status:        { type: string, enum: [pending, received, accepted] }
        sortOrder:     { type: integer }
    RfqQuoteCreate:
      type: object
      required: [factoryName, unitPriceUsd, leadTimeDays, moq]
      properties:
        supplierId:   { type: integer }
        factoryName:  { type: string }
        country:      { type: string }
        unitPriceUsd: { type: number }
        leadTimeDays: { type: integer }
        moq:          { type: integer }
        notes:        { type: string }
        status:       { type: string, enum: [pending, received, accepted] }
    RfqQuoteUpdate:
      type: object
      properties:
        factoryName:  { type: string }
        country:      { type: string }
        unitPriceUsd: { type: number }
        leadTimeDays: { type: integer }
        moq:          { type: integer }
        notes:        { type: ["string", "null"] }
        status:       { type: string, enum: [pending, received, accepted] }
    RfqWithQuotes:
      type: object
      required: [id, product, category, buyerName, targetPriceUsd, quantity, deadline, status, createdAt, quotes]
      properties:
        id:             { type: integer }
        product:        { type: string }
        category:       { type: string }
        buyerName:      { type: string }
        targetPriceUsd: { type: number }
        quantity:       { type: integer }
        deadline:       { type: string, format: date-time }
        status:         { type: string, enum: [open, accepted, cancelled] }
        notes:          { type: ["string", "null"] }
        convertedShipmentId: { type: ["integer", "null"], description: "Shipment ID created when this RFQ was converted to a PO" }
        createdAt:      { type: string, format: date-time }
        quotes:
          type: array
          items: { $ref: "#/components/schemas/RfqQuote" }
    RfqCreate:
      type: object
      required: [product, buyerName, targetPriceUsd, quantity, deadline]
      properties:
        product:        { type: string }
        category:       { type: string }
        buyerName:      { type: string }
        targetPriceUsd: { type: number }
        quantity:       { type: integer }
        deadline:       { type: string, format: date-time }
        notes:          { type: string }
    RfqUpdate:
      type: object
      properties:
        product:        { type: string }
        category:       { type: string }
        buyerName:      { type: string }
        targetPriceUsd: { type: number }
        quantity:       { type: integer }
        deadline:       { type: string, format: date-time }
        notes:          { type: ["string", "null"] }
        status:         { type: string, enum: [open, accepted, cancelled] }
    RfqConvertInput:
      type: object
      required: [acceptedQuoteId, poNumber, supplierId, dueDate, exFactoryDate, destination]
      properties:
        acceptedQuoteId: { type: integer, description: "The RFQ quote ID to mark as accepted" }
        poNumber:        { type: string, description: "Supplier-facing PO number (trader→supplier)" }
        supplierId:      { type: integer }
        dueDate:         { type: string, format: date-time }
        exFactoryDate:   { type: string, format: date-time }
        destination:     { type: string }
        via:             { type: string }
        notes:           { type: string }
        depositPct:      { type: integer, description: "Deposit percentage (default 30)" }
    ProformaInvoice:
      type: object
      required: [rfqId, shipmentId, poNumber, buyerName, product, quantity, unitPriceUsd, totalUsd, depositPct, depositUsd, balanceUsd, deadline, generatedAt]
      properties:
        rfqId:        { type: integer }
        shipmentId:   { type: integer }
        poNumber:     { type: string }
        buyerName:    { type: string }
        product:      { type: string }
        quantity:     { type: integer }
        unitPriceUsd: { type: number }
        totalUsd:     { type: number }
        depositPct:   { type: integer }
        depositUsd:   { type: number }
        balanceUsd:   { type: number }
        paymentTerms: { type: string }
        deadline:     { type: string, format: date-time }
        supplierName: { type: string }
        supplierCountry: { type: string }
        generatedAt:  { type: string, format: date-time }
    SendRfqEmailBody:
      type: object
      required: [to, subject, body]
      properties:
        to:
          type: array
          items: { type: string }
        subject: { type: string }
        body: { type: string }
    SendRfqEmailResponse:
      type: object
      required: [ok]
      properties:
        ok: { type: boolean }